refactoring, hooks file
