+++ tests/features/fake_publisher_client.py
@@ -1,0 +1,19 @@
+import base64
+
+from google.cloud.pubsub_v1 import PublisherClient
+
+import main
+
+
+class FakePublisherClient(PublisherClient):
+    def __init__(self):
+        self.published_messages = []
+
+    def publish(self, topic_path, data):
+        message = {"topic_path": topic_path, "data": data}
+        self.published_messages.append(message)
+
+    def run_all(self):
+        for message in self.published_messages:
+            event = {"data": base64.encodebytes(message["data"])}
+            main.processor(event, {})
